Obtain Cohen's d from the point biserial correlation r

This function converts the point biserial correlation r to Cohen's d.

d_from_r_pb(r_pb, n1, n2, baseRateSensitive = FALSE,
  biasCorrect = FALSE, crosssectionalSampling = FALSE)

Arguments

r_pb
A numerical vector with one or more point biserial r values.

n1, n2
A numerical vector with the sample sizes of the two groups formed by the dichotomous variable. Note that the nth element of these vectors must correspond to the nth element of the r_pbs vector.

baseRateSensitive
Whether to compute tyhe base rate sensitive Cohen's d or not (see McGrath & Meyer, 2006).

biasCorrect
Logical to indicate if the d-values should be bias-corrected. Can also be a vector.

Value

A data frame with in the first column, Cohen's d values, and in the second column, the corresponding variances.

Details

The formula that is used is the following (see e.g. Borenstein et al., 2009):

$$d= \frac{2 r_{pbs}}{\sqrt{1 - r^2}} $$

References

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009) Introduction to Meta-Analysis, Chichester, UK: John Wiley & Sons, Ltd.

Also see:

McGrath, R. E. & Meyer, G. J. (2006) When Effect Sizes Disagree: The Case of r and d. Psychological Methods, 11, 386-401, doi:doi: 10.1037/1082-989X.11.4.386386

Examples

escalc::d_from_r_pb(r_pb = .3,
                    n = 100,
                    proportion=.5);

#> Error in escalc::d_from_r_pb(r_pb = 0.3, n = 100, proportion = 0.5): argument 2 matches multiple formal arguments