Obtain Cohen's d from Student's t

This function converts Student's t to Cohen's d.

d_from_t(t, df, n1, n2, proportion = 0.5)

Arguments

t
The t value.

df
The degrees of freedom of t; either provide this proportion, or n1 and n1.

n1, n2
The sample sizes of the two groups formed by the dichotomous variable.

proportion
The proportion of participants in the first (or second) group; must be specified if df is specified.

Value

A numeric vector of Cohen's d values.

Details

The formula that is used is the following (see e.g. Lakens, 2013):

$$d= t \sqrt{(\frac{1}{n_1} + \frac{1}{n_2})}$$

References

Lakens, D. (2013) Calculating and reporting effect sizes to facilitate cumulative science: a practical primer for t-tests and ANOVAs. Frontiers in Psychology, 4, p. 863. doi: 10.3389/fpsyg.2013.00863

Examples

escalc::d_from_t(t = 2.85, df = 128);

#> [1] 0.4999231